+++ src/app/slice.ts
@@ -10,7 +10,7 @@ import type {
 } from 'sora-js-sdk'
 
 import packageJSON from '../../package.json'
-import { WORKER_SCRIPT } from '../constants'
+import { WORKER_SCRIPT } from '../constants.ts'
 import type {
   AlertMessage,
   DataChannelMessage,
@@ -22,7 +22,7 @@ import type {
   SignalingMessage,
   SoraDevtoolsState,
   TimelineMessage,
-} from '../types'
+} from '../types.ts'
 
 const initialState: SoraDevtoolsState = {
   alertMessages: [],
@@ -255,7 +255,7 @@ export const slice = createSlice({
       const volume = Number.parseFloat(action.payload)
       if (Number.isNaN(volume)) {
         state.fakeVolume = '0'
-      } else if (1 < volume) {
+      } else if (volume > 1) {
         state.fakeVolume = '1'
       } else {
         state.fakeVolume = String(volume)
@@ -461,7 +461,7 @@ export const slice = createSlice({
         title: 'Sora info',
         type: 'info',
         message: action.payload,
-        timestamp: new Date().getTime(),
+        timestamp: Date.now(),
       }
       setAlertMessagesAndLogMessages(state.alertMessages, state.logMessages, alertMessage)
     },
@@ -470,7 +470,7 @@ export const slice = createSlice({
         title: 'Sora error',
         type: 'error',
         message: action.payload,
-        timestamp: new Date().getTime(),
+        timestamp: Date.now(),
       }
       setAlertMessagesAndLogMessages(state.alertMessages, state.logMessages, alertMessage)
     },
@@ -479,7 +479,7 @@ export const slice = createSlice({
         title: 'API info',
         type: 'info',
         message: action.payload,
-        timestamp: new Date().getTime(),
+        timestamp: Date.now(),
       }
       setAlertMessagesAndLogMessages(state.alertMessages, state.logMessages, alertMessage)
     },
@@ -488,7 +488,7 @@ export const slice = createSlice({
         title: 'API error',
         type: 'error',
         message: action.payload,
-        timestamp: new Date().getTime(),
+        timestamp: Date.now(),
       }
       setAlertMessagesAndLogMessages(state.alertMessages, state.logMessages, alertMessage)
     },
@@ -510,7 +510,7 @@ export const slice = createSlice({
     },
     setLogMessages: (state, action: PayloadAction<LogMessage['message']>) => {
       state.logMessages.push({
-        timestamp: new Date().getTime(),
+        timestamp: Date.now(),
         message: {
           title: action.payload.title,
           description: action.payload.description,
@@ -639,7 +639,7 @@ function setAlertMessagesAndLogMessages(
   logMessages: SoraDevtoolsState['logMessages'],
   alertMessage: AlertMessage,
 ): void {
-  if (10 <= alertMessages.length) {
+  if (alertMessages.length >= 10) {
     for (let i = 0; i <= alertMessages.length - 5; i++) {
       alertMessages.pop()
     }