+++ src/components/DevtoolsPane/MetadataForm.tsx
@@ -5,7 +5,7 @@ import { setEnabledMetadata, setMetadata } from '@/app/actions'
 import { useAppDispatch, useAppSelector } from '@/app/hooks'
 import { isFormDisabled } from '@/utils'
 
-import { TooltipFormCheck } from './TooltipFormCheck'
+import { TooltipFormCheck } from './TooltipFormCheck.tsx'
 
 export const MetadataForm: React.FC = () => {
   const enabledMetadata = useAppSelector((state) => state.enabledMetadata)