+++ src/components/DevtoolsPane/SignalingNotifyMetadataForm.tsx
@@ -5,7 +5,7 @@ import { setEnabledSignalingNotifyMetadata, setSignalingNotifyMetadata } from '@
 import { useAppDispatch, useAppSelector } from '@/app/hooks'
 import { isFormDisabled } from '@/utils'
 
-import { TooltipFormCheck } from './TooltipFormCheck'
+import { TooltipFormCheck } from './TooltipFormCheck.tsx'
 
 export const SignalingNotifyMetadataForm: React.FC = () => {
   const enabledSignalingNotifyMetadata = useAppSelector(